# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AeroParts Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscription:
    get:
      operationId: getSubscription
      tags: [subscription]
      summary: Get current seller's subscription info
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionInfo"
        "401":
          description: Unauthorized
...
  /subscription/checkout:
    post:
      operationId: createCheckoutSession
      tags: [subscription]
      summary: Create a Stripe Checkout session for a given price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionInput"
      responses:
        "200":
          description: OK — redirect the browser to the returned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: Missing or invalid priceId
        "401":
          description: Unauthorized
  /subscription/portal:
    post:
      operationId: createPortalSession
      tags: [subscription]
      summary: Create a Stripe Billing Portal session for subscription management
      responses:
        "200":
          description: OK — redirect the browser to the returned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSessionResponse"
        "400":
          description: No billing account found
        "401":
          description: Unauthorized
  /subscription/cancel:
    post:
      operationId: cancelSubscription
      tags: [subscription]
      summary: Cancel the active subscription at period end
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: No active subscription
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    SubscriptionInfo:
      type: object
      required: [plan, effectivePlan, activeListings, listingLimit, canAddListing]
      properties:
        plan:
          type: string
          enum: [free, pro, enterprise, mro_verified, mro_premium]
          description: Subscribed plan (may differ from effectivePlan if payment is lapsed)
        effectivePlan:
          type: string
          enum: [free, pro, enterprise, mro_verified, mro_premium]
          description: Actual enforced plan based on payment status and grace period
        subscriptionStatus:
          type: ["string", "null"]
          enum: [active, trial, past_due, cancelled, suspended, null]
        currentPeriodEnd:
          type: ["string", "null"]
        gracePeriodEnd:
          type: ["string", "null"]
          description: "7 days after currentPeriodEnd when payment is past_due"
        trialEndsAt:
          type: ["string", "null"]
        daysUntilGraceExpires:
          type: ["integer", "null"]
          description: Days remaining in grace period; null if not in grace period
        planExpiresAt:
          type: ["string", "null"]
        activeListings:
          type: integer
        listingLimit:
          type: ["integer", "null"]
        canAddListing:
          type: boolean
        hasFullRfqAccess:
          type: boolean
        mroServiceLimit:
          type: ["integer", "null"]
    CheckoutSessionInput:
      type: object
      required: [priceId]
      properties:
        priceId:
          type: string
          description: Stripe price ID (e.g. price_1ABC...)
    CheckoutSessionResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Stripe Checkout URL — redirect the browser here
    PortalSessionResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Stripe Billing Portal URL — redirect the browser here
...