# lib/api-zod/src/generated/types/checkoutSessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AeroParts Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutSessionInput {
  /** Stripe price ID (e.g. price_1ABC...) */
  priceId: string;
}